+++ test-runner/download-from-vm.js
@@ -20,6 +20,7 @@ if (fs.existsSync(CONFIG_FILE)) {
 
 REMOTE_USERNAME = null
 REMOTE_PASSWORD = null
+REMOTE_HOST = null
 
 const option = {
   host: REMOTE_HOST,