+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change Log
 
-## Unrelease
+## Unrelease `0.14.0`
 
 - **Added**
   - `#198`: As scp user, I want to opt for a prompt if `username` is missing. (`option.prompt`)
   - `#201`: As scp user, I want to opt for a prompt if `password` is missing.
+  - `#204`: As scp user, I want to opt for a prompt if `host` is missing.
 
 ## `0.13.1` - 2021-Aug-4